backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend API"
authors = ["Mustafa Abdallah <mustafa.abdulluh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
django-storages = "^1.13.1"
google-cloud-storage = "^2.6.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
